Rewrite the project to use nlohmann's json lib.
Using nlohmann's json lib we can use MessagePack as our schema for
data organization and that way we can have it parsed with standard json
parsers. Dumped binn because it never worked right and was hard to use.
MessagePack is supported in C and C++ so the client can use a different
MessagePack lib for sending data. I better documented the package api in
the packet processing source.

Rewrote a lot of how CMake works and detects stuff, we now properly
detect a lot of things we didn't before (or did poor detection of stuff)
and also cleaned out a lot of crufty code that made no sense. By default
cmake does "Release" builds and has to be manually told to make debug
builds.